adds game stats module and test
