+++ spec/stat_tracker_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe StatTracker do
 
   stat_tracker = StatTracker.new(locations)
   stat_tracker = StatTracker.from_csv(locations)
-    
+
     it 'exists' do
       expect(stat_tracker).to be_a StatTracker
     end
@@ -29,65 +29,7 @@ RSpec.describe StatTracker do
     end
   end
 
-  context "Game stat methods" do
-    game_path = './spec/fixture_files/test_games.csv'
-    team_path = './spec/fixture_files/test_teams.csv'
-    game_teams_path = './spec/fixture_files/test_game_teams.csv'
-
-    locations = {
-    games: game_path,
-    teams: team_path,
-    game_teams: game_teams_path
-    }
-
-    stat_tracker = StatTracker.from_csv(locations)
-
-    it '#highest_total_score' do
-      expect(stat_tracker.highest_total_score).to eq(9)
-    end
-
-    it '#lowest_total_score' do
-      expect(stat_tracker.lowest_total_score).to eq(1)
-    end
-
-    it '#percentage_home_wins' do
-      expect(stat_tracker.percentage_home_wins).to eq(0.45)
-    end
-
-    it '#percentage_visitor_wins' do
-      expect(stat_tracker.percentage_visitor_wins).to eq(0.37)
-    end
-
-    it '#percentage_ties' do
-      expect(stat_tracker.percentage_ties).to eq(0.18)
-    end
-
-    it '#count_of_games_by_season' do
-      expect(stat_tracker.count_of_games_by_season).to eq({
-        "20122013"=>111,
-        "20132014"=>123,
-        "20142015"=>221,
-        "20152016"=>87,
-        "20162017"=>73,
-        "20172018"=>135
-        })
-    end
-
-    it '#average_goals_per_game' do
-      expect(stat_tracker.average_goals_per_game).to eq(4.2)
-    end
-
-    it '#average_goals_by_season' do
-      expect(stat_tracker.average_goals_by_season).to eq({
-        "20122013" => 4.07,
-        "20162017" => 4.36,
-        "20142015" => 4.07,
-        "20152016" => 4.17,
-        "20132014" => 4.27,
-        "20172018" => 4.39
-        })
-    end
-  end
+  
 
   context 'league stats methods' do
     game_path = './spec/fixture_files/test_games.csv'
@@ -99,7 +41,7 @@ RSpec.describe StatTracker do
       game_teams: game_teams_path
     }
     stat_tracker = StatTracker.from_csv(locations)
-    
+
     it '#count_of_teams' do
       expect(stat_tracker.count_of_teams).to eq(32)
     end
@@ -157,7 +99,7 @@ RSpec.describe StatTracker do
       game_teams: game_teams_path
     }
     stat_tracker = StatTracker.from_csv(locations)
-    
+
     it '#winningest_coach' do
       expect(stat_tracker.winningest_coach("20132014")).to eq("Claude Noel")
     end
@@ -231,11 +173,11 @@ RSpec.describe StatTracker do
 
     it "#find_win_count" do
       expect(stat_tracker.find_win_count("6")).to eq({
-        "20122013"=>[11, 10], 
-        "20172018"=>[16, 10], 
-        "20132014"=>[8, 6], 
-        "20142015"=>[10, 4], 
-        "20152016"=>[3, 1], 
+        "20122013"=>[11, 10],
+        "20172018"=>[16, 10],
+        "20132014"=>[8, 6],
+        "20142015"=>[10, 4],
+        "20152016"=>[3, 1],
         "20162017"=>[2, 0]
         })
     end
